Refactor conftest files
Treat conftest in linkedevents as the place common settings.
Remove obsolete conftest from notifications.

Refs LINK-1122
