+++ events/tests/conftest.py
@@ -1,13 +1,10 @@
 from datetime import datetime, timedelta
 
 # 3rd party
-import pytest
-from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib.gis.geos import MultiPolygon, Point, Polygon
 
 # django
-from django.core.management import call_command
 from django.utils import timezone
 from django_orghierarchy.models import Organization
 from munigeo.models import (
@@ -31,6 +28,7 @@ from events.models import (
     Offer,
     Place,
 )
+from linkedevents.tests.conftest import *  # noqa
 from registrations.models import Registration
 
 from ..models import License, PublicationStatus
@@ -44,22 +42,6 @@ URL = "http://localhost"
 DATETIME = (timezone.now() + timedelta(days=1)).isoformat().replace("+00:00", "Z")
 
 OTHER_DATA_SOURCE_ID = "testotherdatasourceid"
-
-
-# Django test harness tries to serialize DB in order to support transactions
-# within tests. (It restores the snapshot after such tests).
-# This fails with modeltranslate, as the serialization is done before
-# sync_translation_fields has a chance to run. Thus the fields are missing
-# and serialization fails horribly.
-@pytest.fixture(scope="session")
-def django_db_modify_db_settings(django_db_modify_db_settings_xdist_suffix):
-    settings.DATABASES["default"]["TEST"]["SERIALIZE"] = False
-
-
-@pytest.fixture(scope="session")
-def django_db_setup(django_db_setup, django_db_blocker):
-    with django_db_blocker.unblock():
-        call_command("sync_translation_fields", "--noinput")
 
 
 @pytest.fixture
