--- notifications/tests/conftest.py
@@ -1,15 +1,0 @@
-from events.tests.conftest import event  # noqa
-from events.tests.conftest import (
-    administrative_division,
-    administrative_division_type,
-    data_source,
-    django_db_modify_db_settings,
-    django_db_setup,
-    location_id,
-    minimal_event_dict,
-    municipality,
-    organization,
-    place,
-    user,
-    user_api_client,
-)
